New file functions can now set a filename/path directly
Aka when you make a new file in whatever program uses libsyntyche, you
can set the filename at the same time you make the new file. This means
you don't have to first make a new file in the program and then save it
with the name you want.
Note that the code is pretty ugly here and there though, to keep
backward compatibility.